pyproject.toml
[project]
name = "hp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "jinja2>=3.1.6",
    "pyyaml>=6.0.3",
]
